Calculate Incremental Cost-Effectiveness Ratios (ICERs)

This function calculates the Incremental Cost-Effectiveness Ratios (ICERs) for a set of strategies based on their costs and effects.

Usage

calculate_icers(rewards_summary, col_names = c("cost", "utility"))

Arguments

rewards_summary

A matrix or data frame containing the summary statistics of the rewards. It must have columns for cost and utility.

col_names

A character vector specifying the names of the columns for cost and utility. Default is c("cost", "utility").

Value

A data frame with the following columns:

decision: The name of the strategy.

cost: The cost of the strategy.

utility: The effect (utility) of the strategy.

inc_cost: The incremental cost compared to the next less effective strategy.

inc_utility: The incremental effect compared to the next less effective strategy.

ICER: The Incremental Cost-Effectiveness Ratio.

status: The dominance status of the strategy (ND = non-dominated, D = dominated, ED = extendedly dominated).

Examples

rewards_summary <- matrix(c(100, 200, 0.5, 0.7), ncol = 2, 
                          dimnames = list(c("Strategy A", "Strategy B"), 
                                          c("cost", "utility")))
calculate_icers(rewards_summary)
#>              decision cost utility inc_cost inc_utility ICER status
#> Strategy A Strategy A  100     0.5       NA          NA   NA     ND
#> Strategy B Strategy B  200     0.7      100         0.2  500     ND