Run a twig model

This function runs a twig model, which can be either a decision tree or a Markov model.

run_twig(
  twig_obj,
  params,
  n_cycles = NULL,
  verbose = FALSE,
  parallel = FALSE,
  offset_trace_cycle = 1,
  ncore = NULL,
  progress_bar = TRUE
)

Arguments

twig_obj

A twig object created by the twig function.

params

A data frame or list of parameters to be used in the model.

n_cycles

An integer specifying the number of cycles for a Markov model. Default is NULL.

verbose

A logical value indicating whether to print detailed output. Default is FALSE.

parallel

A logical value indicating whether to run the model in parallel. Default is FALSE.

offset_trace_cycle

An integer specifying the offset trace cycle. Default is 1. This is used to adjust the cycle number in the trace output. If set to 0, the initial state distribution will be used as the first cycle. If set to 1, the initial state distribution will be ignored in the Markov trace. In both situations, the total number of cycles will be the same as the input n_cycles.

ncore

An integer specifying the number of cores to use for parallel processing. Default is total number of cores - 1.

progress_bar

A logical value indicating whether to display a progress bar. Default is TRUE.

Value

A list containing the results of the model run. The list includes the following elements:

Rewards_summary An array containing the summary statistics of the rewards if params is a data.frame with more than 1 row.

Rewards_sim An array containing the simulated rewards by simulation.

The following will be included if verbose is TRUE and the twig object is a Markov model:

sim The simulation ID.

Rewards_sim A matrix of simulated rewards of size decision x rewards.

Rewards_array An array containing the rewards for each cycle, state, decision, and reward.

Rewards_array_cycle: An array containing the rewards for each cycle, state, decision, and reward, adjusted for the cycle.

Trace_array: An array containing the Markov trace for each cycle, state, and decision.

TransitionProb_array: An array containing the transition probabilities for each state, destination state, cycle, and decision.

Paths: A data frame containing the path probabilities for each state, cycle, and decision.

Event_options: A data frame containing the event options for each state, cycle, and decision.

Function_Values: A data frame containing the evaluated function values for each state, cycle, decision, and event.

evaluated_funs: A list of evaluated functions. Each function returns a data frame enumerating the dependencies of the function along with the value returned by that function for each combination of values.

path_events: A data frame of the event options along each path. Rows = paths, columns = events.

The following will be included if verbose is TRUE and the twig object is a decision tree:

sim: The simulation ID.

Rewards_sim: A matrix of simulated rewards of size decision x rewards.

path_events: A data frame of the event options along each path. Rows = paths, columns = events.

Event_options: A data frame containing the probability of event options by decision.

evaluated_funs: A list of evaluated functions. Each function returns a data frame enumerating the dependencies of the function along with the value returned by that function for each combination of values.

Prob_Function_Values: A data frame of the probability function values evaluated by decision and events harmonized to the same combinations of decisions and events across all probability functions.

Paths: An array containing the path probabilities of size decision x paths.

Path_rewards: An array containing the path rewards of size decision x paths x rewards. Paths are indexed by their final outcomes in the twig and a key to event options is provided in path_events.

Path_rewards_weighted: An array containing the path rewards weighted by the probability of each path. This is also of size decision x paths x rewards. Paths are indexed by their final outcomes in the twig and a key to event options is provided in path_events.

Outcomes: A matrix of final outcome probabilities. Final outcomes are the terminal event transitions of size decision x outcomes.

See also

Examples

if (FALSE) {
run_twig(mytwig, params, n_cycles = 10)
# see the vignettes for more examples
}